Refactor storage/json package into a more specific domain based one: storage/report
